+++ src/twisted/mail/interfaces.py
@@ -93,7 +93,7 @@ class IServerFactory(Interface):
 
 
 
-class IMailbox(Interface):
+class IMailboxSMTP(Interface):
     """
     An interface for mailbox access.
 
@@ -200,7 +200,7 @@ class IDomain(Interface):
         @type user: L{User}
         @param user: A user.
 
-        @rtype: no-argument callable which returns L{IMessage} provider
+        @rtype: no-argument callable which returns L{IMessageSMTPI} provider
         @return: A function which takes no arguments and returns a message
             receiver for the user.
 
@@ -231,6 +231,20 @@ class IDomain(Interface):
 
 
 
+class IAlias(Interface):
+    """
+    An interface for aliases.
+    """
+    def createMessageReceiver():
+        """
+        Create a message receiver.
+
+        @rtype: L{IMessageSMTP} provider
+        @return: A message receiver.
+        """
+
+
+
 class IAliasableDomain(IDomain):
     """
     An interface for email domains which can be aliased to other domains.
@@ -251,12 +265,13 @@ class IAliasableDomain(IDomain):
         @type user: L{User}
         @param user: A user.
 
-        @type memo: L{None} or L{dict} of L{AliasBase <twisted.mail.mail.AliasBase>}
+        @type memo: L{None} or L{dict} of
+            L{AliasBase <twisted.mail.alias.AliasBase>}
         @param memo: A record of the addresses already considered while
-            resolving aliases.  The default value should be used by all
-            external code.
+            resolving aliases. The default value should be used by all external
+            code.
 
-        @rtype: no-argument callable which returns L{IMessage} provider
+        @rtype: no-argument callable which returns L{IMessageSMTP} provider
         @return: A function which takes no arguments and returns a message
             receiver for the user.
 
@@ -295,13 +310,13 @@ class IMessageDelivery(Interface):
         @param user: The address to validate.
 
         @rtype: no-argument callable
-        @return: A C{Deferred} which becomes, or a callable which
-        takes no arguments and returns an object implementing L{IMessage}.
-        This will be called and the returned object used to deliver the
-        message when it arrives.
+        @return: A C{Deferred} which becomes, or a callable which takes no
+            arguments and returns an object implementing L{IMessageSMTP}. This
+            will be called and the returned object used to deliver the message
+            when it arrives.
 
-        @raise SMTPBadRcpt: Raised if messages to the address are
-        not to be accepted.
+        @raise SMTPBadRcpt: Raised if messages to the address are not to be
+            accepted.
         """
 
     def validateFrom(helo, origin):
@@ -344,7 +359,7 @@ class IMessageDeliveryFactory(Interface):
 
 
 
-class IMessage(Interface):
+class IMessageSMTP(Interface):
     """
     Interface definition for messages that can be sent via SMTP.
     """
@@ -369,4 +384,334 @@ class IMessage(Interface):
         Handle message truncated.
 
         semantics should be to discard the message
+        """
+
+
+class IMessageIMAPPart(Interface):
+    def getHeaders(negate, *names):
+        """Retrieve a group of message headers.
+
+        @type names: C{tuple} of C{str}
+        @param names: The names of the headers to retrieve or omit.
+
+        @type negate: C{bool}
+        @param negate: If True, indicates that the headers listed in C{names}
+        should be omitted from the return value, rather than included.
+
+        @rtype: C{dict}
+        @return: A mapping of header field names to header field values
+        """
+
+    def getBodyFile():
+        """Retrieve a file object containing only the body of this message.
+        """
+
+    def getSize():
+        """Retrieve the total size, in octets, of this message.
+
+        @rtype: C{int}
+        """
+
+    def isMultipart():
+        """Indicate whether this message has subparts.
+
+        @rtype: C{bool}
+        """
+
+    def getSubPart(part):
+        """Retrieve a MIME sub-message
+
+        @type part: C{int}
+        @param part: The number of the part to retrieve, indexed from 0.
+
+        @raise IndexError: Raised if the specified part does not exist.
+        @raise TypeError: Raised if this message is not multipart.
+
+        @rtype: Any object implementing L{IMessageIMAPPart}.
+        @return: The specified sub-part.
+        """
+
+class IMessageIMAP(IMessageIMAPPart):
+    def getUID():
+        """Retrieve the unique identifier associated with this message.
+        """
+
+    def getFlags():
+        """Retrieve the flags associated with this message.
+
+        @rtype: C{iterable}
+        @return: The flags, represented as strings.
+        """
+
+    def getInternalDate():
+        """Retrieve the date internally associated with this message.
+
+        @rtype: C{str}
+        @return: An RFC822-formatted date string.
+        """
+
+class IMessageIMAPFile(Interface):
+    """Optional message interface for representing messages as files.
+
+    If provided by message objects, this interface will be used instead
+    the more complex MIME-based interface.
+    """
+    def open():
+        """Return a file-like object opened for reading.
+
+        Reading from the returned file will return all the bytes
+        of which this message consists.
+        """
+
+class ISearchableIMAPMailbox(Interface):
+    def search(query, uid):
+        """Search for messages that meet the given query criteria.
+
+        If this interface is not implemented by the mailbox, L{IMailboxIMAP.fetch}
+        and various methods of L{IMessageIMAP} will be used instead.
+
+        Implementations which wish to offer better performance than the
+        default implementation should implement this interface.
+
+        @type query: C{list}
+        @param query: The search criteria
+
+        @type uid: C{bool}
+        @param uid: If true, the IDs specified in the query are UIDs;
+        otherwise they are message sequence IDs.
+
+        @rtype: C{list} or C{Deferred}
+        @return: A list of message sequence numbers or message UIDs which
+        match the search criteria or a C{Deferred} whose callback will be
+        invoked with such a list.
+
+        @raise IllegalQueryError: Raised when query is not valid.
+        """
+
+class IMessageIMAPCopier(Interface):
+    def copy(messageObject):
+        """
+        Copy the given message object into this mailbox.
+
+        The message object will be one which was previously returned by
+        L{IMailboxIMAP.fetch}.
+
+        Implementations which wish to offer better performance than the default
+        implementation should implement this interface.
+
+        If this interface is not implemented by the mailbox,
+        L{IMailboxIMAP.addMessage} will be used instead.
+
+        @rtype: C{Deferred} or C{int}
+        @return: Either the UID of the message or a Deferred which fires with
+            the UID when the copy finishes.
+        """
+
+class IMailboxIMAPInfo(Interface):
+    """Interface specifying only the methods required for C{listMailboxes}.
+
+    Implementations can return objects implementing only these methods for
+    return to C{listMailboxes} if it can allow them to operate more
+    efficiently.
+    """
+
+    def getFlags():
+        """Return the flags defined in this mailbox
+
+        Flags with the \\ prefix are reserved for use as system flags.
+
+        @rtype: C{list} of C{str}
+        @return: A list of the flags that can be set on messages in this mailbox.
+        """
+
+    def getHierarchicalDelimiter():
+        """Get the character which delimits namespaces for in this mailbox.
+
+        @rtype: C{str}
+        """
+
+class IMailboxIMAP(IMailboxIMAPInfo):
+    def getUIDValidity():
+        """Return the unique validity identifier for this mailbox.
+
+        @rtype: C{int}
+        """
+
+    def getUIDNext():
+        """Return the likely UID for the next message added to this mailbox.
+
+        @rtype: C{int}
+        """
+
+    def getUID(message):
+        """Return the UID of a message in the mailbox
+
+        @type message: C{int}
+        @param message: The message sequence number
+
+        @rtype: C{int}
+        @return: The UID of the message.
+        """
+
+    def getMessageCount():
+        """Return the number of messages in this mailbox.
+
+        @rtype: C{int}
+        """
+
+    def getRecentCount():
+        """Return the number of messages with the 'Recent' flag.
+
+        @rtype: C{int}
+        """
+
+    def getUnseenCount():
+        """Return the number of messages with the 'Unseen' flag.
+
+        @rtype: C{int}
+        """
+
+    def isWriteable():
+        """Get the read/write status of the mailbox.
+
+        @rtype: C{int}
+        @return: A true value if write permission is allowed, a false value otherwise.
+        """
+
+    def destroy():
+        """Called before this mailbox is deleted, permanently.
+
+        If necessary, all resources held by this mailbox should be cleaned
+        up here.  This function _must_ set the \\Noselect flag on this
+        mailbox.
+        """
+
+    def requestStatus(names):
+        """Return status information about this mailbox.
+
+        Mailboxes which do not intend to do any special processing to
+        generate the return value, C{statusRequestHelper} can be used
+        to build the dictionary by calling the other interface methods
+        which return the data for each name.
+
+        @type names: Any iterable
+        @param names: The status names to return information regarding.
+        The possible values for each name are: MESSAGES, RECENT, UIDNEXT,
+        UIDVALIDITY, UNSEEN.
+
+        @rtype: C{dict} or C{Deferred}
+        @return: A dictionary containing status information about the
+        requested names is returned.  If the process of looking this
+        information up would be costly, a deferred whose callback will
+        eventually be passed this dictionary is returned instead.
+        """
+
+    def addListener(listener):
+        """Add a mailbox change listener
+
+        @type listener: Any object which implements C{IMailboxListener}
+        @param listener: An object to add to the set of those which will
+        be notified when the contents of this mailbox change.
+        """
+
+    def removeListener(listener):
+        """Remove a mailbox change listener
+
+        @type listener: Any object previously added to and not removed from
+        this mailbox as a listener.
+        @param listener: The object to remove from the set of listeners.
+
+        @raise ValueError: Raised when the given object is not a listener for
+        this mailbox.
+        """
+
+    def addMessage(message, flags = (), date = None):
+        """Add the given message to this mailbox.
+
+        @type message: A file-like object
+        @param message: The RFC822 formatted message
+
+        @type flags: Any iterable of C{str}
+        @param flags: The flags to associate with this message
+
+        @type date: C{str}
+        @param date: If specified, the date to associate with this
+        message.
+
+        @rtype: C{Deferred}
+        @return: A deferred whose callback is invoked with the message
+        id if the message is added successfully and whose errback is
+        invoked otherwise.
+
+        @raise ReadOnlyMailbox: Raised if this Mailbox is not open for
+        read-write.
+        """
+
+    def expunge():
+        """Remove all messages flagged \\Deleted.
+
+        @rtype: C{list} or C{Deferred}
+        @return: The list of message sequence numbers which were deleted,
+        or a C{Deferred} whose callback will be invoked with such a list.
+
+        @raise ReadOnlyMailbox: Raised if this Mailbox is not open for
+        read-write.
+        """
+
+    def fetch(messages, uid):
+        """Retrieve one or more messages.
+
+        @type messages: C{MessageSet}
+        @param messages: The identifiers of messages to retrieve information
+        about
+
+        @type uid: C{bool}
+        @param uid: If true, the IDs specified in the query are UIDs;
+        otherwise they are message sequence IDs.
+
+        @rtype: Any iterable of two-tuples of message sequence numbers and
+        implementors of C{IMessageIMAP}.
+        """
+
+    def store(messages, flags, mode, uid):
+        """Set the flags of one or more messages.
+
+        @type messages: A MessageSet object with the list of messages requested
+        @param messages: The identifiers of the messages to set the flags of.
+
+        @type flags: sequence of C{str}
+        @param flags: The flags to set, unset, or add.
+
+        @type mode: -1, 0, or 1
+        @param mode: If mode is -1, these flags should be removed from the
+        specified messages.  If mode is 1, these flags should be added to
+        the specified messages.  If mode is 0, all existing flags should be
+        cleared and these flags should be added.
+
+        @type uid: C{bool}
+        @param uid: If true, the IDs specified in the query are UIDs;
+        otherwise they are message sequence IDs.
+
+        @rtype: C{dict} or C{Deferred}
+        @return: A C{dict} mapping message sequence numbers to sequences of C{str}
+        representing the flags set on the message after this operation has
+        been performed, or a C{Deferred} whose callback will be invoked with
+        such a C{dict}.
+
+        @raise ReadOnlyMailbox: Raised if this mailbox is not open for
+        read-write.
+        """
+
+class ICloseableMailboxIMAP(Interface):
+    """A supplementary interface for mailboxes which require cleanup on close.
+
+    Implementing this interface is optional.  If it is implemented, the protocol
+    code will call the close method defined whenever a mailbox is closed.
+    """
+    def close():
+        """Close this mailbox.
+
+        @return: A C{Deferred} which fires when this mailbox
+        has been closed, or None if the mailbox can be closed
+        immediately.
         """
